pyproject.toml
[project]
name = "neurips-sessions"
version = "0.1.0"
description = "Capture slides, video, and AI-generated notes from NeurIPS sessions"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "yt-dlp>=2024.1.1",
    "myslideslive>=0.1.0",
    "faster-whisper>=1.0.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.27.0",
]

[project.scripts]
neurips-capture = "capture:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["*.py"]

[tool.ruff]
line-length = 100
